+++ src/main/webapp/entries.js
@@ -9,7 +9,7 @@ module.exports = {
     "regenerator-runtime/runtime",
     "expose-loader?$!jquery",
     "expose-loader?angular!angular",
-    "./resources/js/vendors"
+    "./resources/js/vendors",
   ],
   login: "./resources/js/pages/LoginPage.jsx",
   access_confirmation: "./resources/js/pages/oauth/access_confirmation.js",
@@ -46,6 +46,7 @@ module.exports = {
     "./resources/js/pages/projects/settings/project-settings-basic.js",
   "project-settings-remote":
     "./resources/js/pages/projects/settings/project-settings-remote.js",
+  "project-details": "./resources/js/pages/projects/details/index.js",
   "project-sync": "./resources/js/pages/projects/project-sync.js",
   "remote-api-details":
     "./resources/js/pages/remote-apis/remote-api-details.js",
@@ -83,5 +84,5 @@ module.exports = {
   "ncbi-exports": "./resources/js/pages/projects/ncbi-export.js",
   search: "./resources/js/pages/search/search.js",
   "run-files": "./resources/js/pages/sequence-files/run-files.js",
-  "user-details": "./resources/js/pages/user-details.js"
+  "user-details": "./resources/js/pages/user-details.js",
 };