+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Changes
 * [UI]: Added selection to export assemblies to Galaxy.
 * [UI]: Updated Project > Members page to remove `DataDables` and use Ant Design.
 * [UI/Developer]: Added image file viewer to analysis outputs view.
+* [UI]: Moved project > details into the project > settings panel.
 
 19.09 to 20.01
 --------------